src/main/java/com/interview/creditcard/validation/SecurityNumberValidator.java
package com.interview.creditcard.validation;

import javax.validation.Constraint;
import javax.validation.ConstraintValidator;
import javax.validation.ConstraintValidatorContext;

public class SecurityNumberValidator implements ConstraintValidator<SecurityCodeValidator, String> {
    @Override
    public boolean isValid(String s, ConstraintValidatorContext constraintValidatorContext) {
        if (s.isEmpty() || s.length() != 3)
            return false;
        try {
            Integer.parseInt(s);
        } catch (NumberFormatException n) {
            return false;
        }
        return true;
    }
}
